Working V1 of the package
Includes tests, and travis-ci integration.

HTML POST working example in examples/poc
